working on e-wallet

+++ lib/model/getMyWallteResponse.dart/getMyWalletResponse.dart
@@ -1,29 +1,32 @@
 import 'dart:convert';
 
-GetMyWalletResponse getMyWalletResponseFromJson(String str) => GetMyWalletResponse.fromJson(json.decode(str));
+GetMyWalletResponse getMyWalletResponseFromJson(String str) =>
+    GetMyWalletResponse.fromJson(json.decode(str));
 
-String getMyWalletResponseToJson(GetMyWalletResponse data) => json.encode(data.toJson());
+String getMyWalletResponseToJson(GetMyWalletResponse data) =>
+    json.encode(data.toJson());
 
 class GetMyWalletResponse {
-    GetMyWalletResponse({
-        this.statusCode,
-        this.message,
-        this.data,
-    });
+  GetMyWalletResponse({
+    this.statusCode,
+    this.message,
+    this.data,
+  });
 
-    int ?statusCode;
-    String ?message;
-    double ?data;
+  int? statusCode;
+  String? message;
+  int? data;
 
-    factory GetMyWalletResponse.fromJson(Map<String, dynamic> json) => GetMyWalletResponse(
+  factory GetMyWalletResponse.fromJson(Map<String, dynamic> json) =>
+      GetMyWalletResponse(
         statusCode: json["StatusCode"],
         message: json["Message"],
         data: json["Data"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "StatusCode": statusCode,
         "Message": message,
         "Data": data,
-    };
+      };
 }
